requirements-dev.txt
# Include production requirements
-r requirements.txt

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.10.1
ruff==0.1.6
mypy==1.7.1
django-stubs==4.2.6
djangorestframework-stubs==3.14.4
ipython==8.17.2
django-debug-toolbar==4.2.0